# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Events API - Luma-Inspired Event Management",
        "description": "Open-source headless Laravel event management API from Maak. Features organizations, events, ticket types, registrations, QR check-ins, waitlist management, and team member invitations.",
        "contact": {
            "name": "Maak",
            "url": "https://github.com/maakwizardry/events"
        },
        "version": "1.1.0"
    },
    "servers": [
        {
            "url": "http://events.test/api/v1",
            "description": "Local Development Server"
        },
        {
            "url": "http://localhost:8000/api/v1",
            "description": "PHP Built-in Server"
        }
    ],
    "paths": {
...
        "/organizations/{organization}/events": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "List organization events",
                "description": "Get all events for a specific organization. Requires user to be a member of the organization.",
                "operationId": "e1c459075da9a9cf1eef4393823aeced",
                "parameters": [
                    {
                        "name": "organization",
                        "in": "path",
                        "description": "Organization UUID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of organization events",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "uuid": {
                                                        "type": "string",
                                                        "example": "550e8400-e29b-41d4-a716-446655440000"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Tech Conference 2024"
                                                    },
                                                    "slug": {
                                                        "type": "string",
                                                        "example": "tech-conference-2024"
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "example": "published"
                                                    },
                                                    "visibility": {
                                                        "type": "string",
                                                        "example": "public"
                                                    },
                                                    "starts_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "registrations_count": {
                                                        "type": "integer",
                                                        "example": 45
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Not a member of this organization"
                    },
                    "404": {
                        "description": "Organization not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/events/{event}": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Get event details",
                "description": "Get detailed information about a specific event. Requires authorization to view the event.",
                "operationId": "8acdb3ac7ee5021ae9834bab337f823b",
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "description": "Event UUID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Event details"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Not authorized to view this event"
                    },
                    "404": {
                        "description": "Event not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Events"
                ],
                "summary": "Update event",
                "description": "Update event details. Requires admin or owner role in the organization.",
                "operationId": "12ed6e275fb359902fb8973e137fd73d",
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "description": "Event UUID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Tech Conference 2024"
                                    },
                                    "slug": {
                                        "type": "string",
                                        "example": "tech-conference-2024"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "visibility": {
                                        "type": "string",
                                        "enum": [
                                            "public",
                                            "private"
                                        ]
                                    },
                                    "location_type": {
                                        "type": "string",
                                        "enum": [
                                            "physical",
                                            "online",
                                            "hybrid"
                                        ]
                                    },
                                    "starts_at": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "ends_at": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "capacity": {
                                        "type": "integer"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "draft",
                                            "published",
                                            "cancelled",
                                            "completed"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Event updated successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Not authorized to update this event"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Events"
                ],
                "summary": "Delete event",
                "description": "Delete an event. Requires admin or owner role in the organization.",
                "operationId": "7b37be6da1e9cce8da63d2ce18813309",
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "description": "Event UUID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Event deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Event deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Not authorized to delete this event"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/events/{event}/availability": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Get event availability (authenticated)",
                "description": "Get event capacity and availability information. Requires authorization to view the event.",
                "operationId": "ba1afb8c643e9c52bb32bffb5c2e2502",
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "description": "Event UUID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Event availability details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "capacity": {
                                            "type": "integer",
                                            "example": 100
                                        },
                                        "total_registered": {
                                            "type": "integer",
                                            "example": 45
                                        },
                                        "available_spots": {
                                            "type": "integer",
                                            "example": 55
                                        },
                                        "is_full": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "enable_waitlist": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "is_registration_open": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "registration_opens_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "nullable": true
                                        },
                                        "registration_closes_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Not authorized to view this event"
                    },
                    "404": {
                        "description": "Event not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Use the token from POST /auth/login or POST /auth/register",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}